Update patch set 1
Patch Set 1:

Say here is gerrit and only you can change the source code of CyanogenMod so I am suggesting is that I am software development student can find a better provider so you can be using for yahoo weather between the two I prefer the accuweather. If I could work deeper in dealing directly with the code would be an honor to collaborate.

Patch-set: 1